+++ react-app/src/App.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
-import LoginForm from './components/auth/LoginForm';
-import SignUpForm from './components/auth/SignUpForm';
+import LoginForm from './components/Starter/auth/LoginForm';
+import SignUpForm from './components/Starter/auth/SignUpForm';
 import NavBar from './components/NavBar';
-import ProtectedRoute from './components/auth/ProtectedRoute';
-import UsersList from './components/UsersList';
-import User from './components/User';
+import ProtectedRoute from './components/Starter/auth/ProtectedRoute';
+import UsersList from './components/Starter/auth/UsersList';
+import User from './components/Starter/auth/User';
 import { authenticate } from './store/session';
 
 function App() {
