+++ react-app/src/components/Starter/NavBar.js
@@ -1,7 +1,7 @@
 
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import LogoutButton from './auth/LogoutButton';
+import LogoutButton from './Starter/auth/LogoutButton';
 
 const NavBar = () => {
   return (
