+++ react-app/src/components/Starter/auth/LoginForm.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Redirect } from 'react-router-dom';
-import { login } from '../../store/session';
+import { login } from '../../../store/session';
 
 const LoginForm = () => {
   const [errors, setErrors] = useState([]);
